Added more plots
Almost done, for now. This is an exhausting dataset due to the number of confusing features (even though the documentation itself is good, I still feel there is redundant information, not to mention an entire second dataset which I still do not understand what it stands for) so will take a break from it for a few days.